[Finished and Tested] Admin
No errors and no memory leaks. All commented

+++ Project/src/admin.c
@@ -4,9 +4,6 @@
 #include <errno.h>
 #include <unistd.h>
 
-#include <sys/un.h>
-#include <sys/socket.h>
-#include <sys/types.h>
 #include <arpa/inet.h>
 
 #include "client-server.pb-c.h"
@@ -20,8 +17,6 @@
 #define LOG 0
 #define QUIT 1
 #define DISC 2
-
-#define STD_SIZE 100
 
 
 //Function Declaration (.h not necessary)
@@ -42,9 +37,18 @@ int main ( int arc, char **argv ) {
 		if (handleKeyboard() != 0) break;
 	}
 	
+	close(sock_fd);
+	printf("\nSee you soon!\n");
 	exit(0);
 }
 
+/* handleKeyboard
+ * 
+ * Handles admin keyboard input
+ * 
+ * 
+ * @ returns int - 1 for exit, 0 for repeat
+ * */
 int handleKeyboard(){
 	int should_exit = 0;
 	char line[STD_SIZE];
@@ -53,24 +57,34 @@ int handleKeyboard(){
 	fgets(line, STD_SIZE, stdin);
 	if(sscanf(line, "%s", command) == 1){
 		if(strcmp(command, LOG_STR) == 0){
-			printf("Sending LOG request to server\n");
+			printf("[Admin] Sending LOG request to server\n");
 			if (sendRequest(LOG) == -1) should_exit = 1;
 			if (receiveLog() == -1) should_exit = 1;
 		}else if(strcmp(command, QUIT_STR)==0){
-			printf("Sending QUIT request to server\n");
+			printf("[Admin] Sending QUIT request to server\n");
 			sendRequest(QUIT);
 			should_exit = 1;
 		}else if(strcmp(command, DISC_STR)==0){
-			printf("Sending DISC request to server\n");
+			printf("[Admin] Sending DISC request to server\n");
 			sendRequest(DISC);
 			should_exit = 1;
+		}else{
+			printf("[Admin] Invalid command\n");
 		}
 	}else{
-		printf("Error in command\n");
+		printf("[Admin] Error in command\n");
 	}
 	return should_exit;
 }
 
+/* sendRequest
+ * 
+ * Receives action desired of the server.
+ * Sends request.
+ * 
+ * @ int action
+ * @ returns int - Failiure or Success
+ * */
 int sendRequest(int action){
 	ADMIN message;
 	admin__init(&message);
@@ -81,20 +95,30 @@ int sendRequest(int action){
 	return sendMessage( disconnect_message , sock_fd);
 }
 
-int receiveLog(){
-	printf("Receiving LOG request from server\n");
-	
+
+/* receiveLog
+ * 
+ * Receives LOG form server and prints to terminal
+ * 
+ * @ returns int - Failiure or Success
+ * */
+int receiveLog(){	
 	proto_msg * message = createProtoMSG( DONT_ALLOC_MSG );
 	message->msg = malloc(LOG_SIZE * (sizeof(char)));
 	message->msg_size = read(sock_fd, message->msg, LOG_SIZE);
 	//Socket closed abruptly
 	if ( message->msg_size == 0){
-		printf("Socket Closed Abruptly\n");
-		close(sock_fd);
+		printf("[System] Error receiving messages from server. Disconnecting...\n");
 		destroyProtoMSG(message);
 		return -1;
 	}
-	ADMIN *admin = admin__unpack(NULL, message->msg_size, message->msg);
+	
+	ADMIN *admin = admin__unpack(NULL, message->msg_size, (const uint8_t *) message->msg);
+	if (admin == NULL){
+		printf("[System] Message from server in incorrect format (Discarted)\n");
+		return 0;
+	}
+	
 	printf("%s", admin->log);
 	destroyProtoMSG(message);
 	admin__free_unpacked(admin, NULL);
@@ -102,20 +126,17 @@ int receiveLog(){
 }
 
 
-
-
 /* iniSocket
  * 
  * Initializes the socket connection
  * Connects to server on PORT 3001, IP 127.0.0.1
  * 
- * @ returns sock_fd - Socket descriptor
  * */
 void iniSocket(){
 	struct sockaddr_in server_addr;
 	sock_fd = socket(AF_INET, SOCK_STREAM, 0);
 	if(sock_fd == -1){
-		perror("socket ");
+		perror("[System Error] Socket ");
 		exit(-1);
 	}
 
@@ -124,7 +145,7 @@ void iniSocket(){
 	inet_aton("127.0.0.1", & server_addr.sin_addr); /*IP*/	
 
 	if( connect(sock_fd, ( struct sockaddr *) &server_addr, sizeof(server_addr)) == -1){
-		perror("connect ");
+		perror("[System Error] Connect ");
 		exit(-1);
 	}
